backend/requirements.txt
fastapi
uvicorn[standard]
python-dotenv
psycopg2-binary
sqlalchemy
httpx
supabase
openai
sendgrid
apscheduler
python-jose[cryptography]
passlib[bcrypt]
python-multipart
pydantic 